feat(kernel): perform elementwise product in FP32

+++ include/kernel.h
@@ -38,4 +38,4 @@ float launch_square_sum_reduce(const Tensor &x);
 __global__ void elementwise_product(__nv_bfloat16 *__restrict__ out,
                                     const __nv_bfloat16 *__restrict__ x,
                                     const __nv_bfloat16 *__restrict__ y,
-                                    __nv_bfloat16 scale, std::size_t n);
+                                    float scale, std::size_t n);
